chore: restructure some codes to separate files
