+++ src/pgwire/CMakeLists.txt
@@ -1,9 +1,11 @@
 add_library(pgwire STATIC
   buffer.cpp
+  exception.cpp
   io.cpp
   log.cpp
   protocol.cpp
   server.cpp
+  session.cpp
   types.cpp
   utils.cpp
   writer.cpp
